refactor: consolidate DLL gateway files under services/gateway/
DllGatewayServer and DllGatewayClient lived at services/ root while
MarketDataGatewayService lived under services/gateway/, even though all
three are PFCF bridge components with the same boundary role (cross-
process access to exchange functionality). Moving the DLL gateway pair
into services/gateway/ completes the convention: everything under
services/gateway/ is now a gateway, and nothing outside it claims to be.

Updates CLAUDE.md path references and appends an update note to
ADR-004 recording the relocation.
