+++ process/run_order_executor_gateway.py
@@ -19,7 +19,7 @@ sys.path.append(str(Path(__file__).resolve().parent))
 
 from src.infrastructure.messaging import ZmqPuller
 from src.domain.order.order_executor_gateway import OrderExecutorGateway
-from src.infrastructure.services.dll_gateway_client import DllGatewayClient
+from src.infrastructure.services.gateway.dll_gateway_client import DllGatewayClient
 from src.infrastructure.repositories.session_json_file_repository import SessionJsonFileRepository
 from src.infrastructure.loggers.logger_default import LoggerDefault
 from src.app.cli_pfcf.config import Config
