+++ src/infrastructure/services/gateway/dll_gateway_client.py
